+++ zeroproof/simulations/__init__.py
@@ -97,6 +97,12 @@ from .score.benchmark import benchmark_report, format_benchmark, run_benchmark
 from .score.curriculum import curriculum, format_curriculum, retire_solved
 from .score.delta import delta_report, format_delta_report
 from .score.grading import behavior_signature, conduct_grade
+from .score.grounding import (
+    argument_grounding,
+    grounding_report,
+    mark_grounding,
+    ungrounded_arguments,
+)
 from .score.hygiene import (
     HACK_THRESHOLD,
     dedupe_groups,
@@ -178,6 +184,7 @@ __all__ = [
     "adaptive_allocator",
     "agents",
     "allocator_slot_counts",
+    "argument_grounding",
     "attach_delta",
     "behavior_signature",
     "behavioral_markers",
@@ -221,6 +228,7 @@ __all__ = [
     "get_run",
     "grade",
     "grade_llm",
+    "grounding_report",
     "group_signal",
     "hf_publish",
     "hf_publish_run",
@@ -239,6 +247,7 @@ __all__ = [
     "local_model",
     "logprob_report",
     "loss_mask",
+    "mark_grounding",
     "mark_rows",
     "marker_summary",
     "mean_kl",
@@ -294,6 +303,7 @@ __all__ = [
     "training_run",
     "trim_out_of_band",
     "trim_unanimous_groups",
+    "ungrounded_arguments",
     "unpublish",
     "update_dataset",
     "validate",